implement commands feature (upgrade and kill)
