+++ config.ini
@@ -5,7 +5,7 @@ group = dfc13_collab
 [Queue]
 n_run = 260
 n_queue = 50
-sleep = 100
+sleep = 10
 
 [Directories]
 basedir = ~/../<!User!>/PBS
@@ -18,3 +18,7 @@ qsub_info = qsub_info
 
 [Logging]
 qstat_cache_dir = /gpfs/group/dfc13/default/qstat_out
+
+[Commands]
+shutdown = False
+upgrade = force
